I fixed the light trail issue #23 and issue #26. The situation is now as followed:
A light trail now holds a list of squares. Each time a new square is
added to the light trail, the trail sets the light trail boolean on that
square to true. If a square is removed from the light trail, because it
was too long, the light trail boolean on that square is reset to false.
This way a square does not know to which the light trail belongs but
this can be fix without much work. So I think this is a good solution
...